requirements.txt
asgiref==3.11.0
Django==4.2.26
django-phonenumber-field==8.3.0
phonenumbers==9.0.19
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
